Provide optional per-column SQLFORM.grid CSS class
I wanted to control the CSS in the TD elements for one column in a SQLFORM.grid. I've added columnclasses={} to the grid() parameters, which is a dictionary of 'tablename.fieldname':'classname' e.g. {'t_mytable.f_myfield' : 'myclass'}. If you accept this pull request, I'll update the book pages (if you want: let me know).