Add a fired event after the component has finished loading
Sometimes we may want to know when the component is finished with it’s
loading process entirely to trigger animations or something similar.  I
added this for my own website, and I think that others may want to use
this.  It’s a small addition, but I hope that it will be helpful to
others!