Added option to exclude unchanged fields from form update
Normally, when a single field is changed in a form, every field in the form is persisted. I assume this is done to ensure record integrity between fields. However this makes it harder to determine which field actually changed.

By adding the option table._update_changed_form_fields_only = True, unchanged fields will now be excluded from the update.